+++ src/app/(auth)/_layout.tsx
@@ -1,19 +1,5 @@
-import { getToken } from '@/lib/storage'
-import { router, Stack, usePathname } from 'expo-router'
-import { useEffect } from 'react'
+import { Stack } from 'expo-router'
 
 export default function Layout() {
-  const pathname = usePathname();
-
-  useEffect(() => {
-    const checkAccess = async () => {
-      const preventNavigate = ['/login', '/login/profile-setup'];
-      const isLoggedIn = await getToken('accessToken')
-      if (!isLoggedIn && !preventNavigate.includes(pathname)) router.replace('/login')
-    }
-
-    checkAccess().catch(console.error)
-  }, [pathname])
-
   return <Stack screenOptions={{ headerShown: false }} />
 }
